requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
exceptiongroup==1.2.2
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
milvus==2.3.9
milvus-lite==2.4.10
numpy==2.0.2
ollama==0.4.4
pandas==2.2.3
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
pymilvus==2.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
